Django Project/cycle/cycle/requirements.txt
django
whitenoise
psycopg2
python-dotenv
django-location-field
pillow
channels
channels_redis
